Replace PreviewHtmlUi$ console application test with MiniServerTest unit test.
